# api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "contact": {
      "email": "developer.overheid@geonovum.nl",
      "name": "Team developer.overheid.nl",
      "url": "https://github.com/developer-overheid-nl/don-tools-api/issues"
    },
    "description": "API to access the tools endpoints of developer.overheid.nl.\n\n## Auth\n\nThis API provides tools endpoints that can be accessed with either an API key or a client credentials token.\n\n### API key\n\nUsing an API key, you can access the tools endpoints.\nThese requests can also be made from the browser.\nRequest a read-only API key at https://apis.developer.overheid.nl/apis/key-aanvragen.\nSimply pass the obtained API key with each request using the `X-Api-Key` header.\n\n### Client credentials token\n\nUsing a client credentials token, you can access the tools endpoints.\nTo obtain the token, perform a `POST` request to `https://auth.developer.overheid.nl/realms/don/protocol/openid-connect/token` with the following Form URL Encoded body:\n- `grant_type`: `client_credentials`\n- `scope`: `tools`\n- `client_id`: the client id you received from us\n- `client_secret`: the client secret you received from us\n\nPass the obtained token with each request using the `Authorization` header. Example:\n\n`Authorization`: `Bearer {ACCESS_TOKEN}` (replace `{ACCESS_TOKEN}` with the obtained `access_token`)",
    "title": "Tools API v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "description": "Production",
      "url": "https://api.developer.overheid.nl/tools/v1"
    },
    {
      "description": "Test",
      "url": "https://api.don.apps.digilab.network/tools/v1"
    }
  ],
  "tags": [
    {
      "description": "Tools API V1 routes",
      "name": "API v1"
    },
    {
      "description": "Conversies en hulpmiddelen",
      "name": "Tools"
    }
  ],
  "paths": {
    "/v1/arazzo/markdown": {
      "post": {
        "description": "Genereert alleen de Markdown-uitvoer van een Arazzo specificatie. Body: { oasUrl|oasBody }",
        "operationId": "arazzoMarkdown",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OasInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "OK",
            "headers": {
              "API-Version": {
                "description": "De API-versie van de response",
                "explode": false,
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "apiKey": [],
            "clientCredentials": []
          }
        ],
        "summary": "Arazzo Markdown (POST)",
        "tags": [
          "Tools"
        ]
      }
    },
    "/v1/arazzo/mermaid": {
      "post": {
        "description": "Genereert alleen de Mermaid flowchart van een Arazzo specificatie. Body: { oasUrl|oasBody }",
        "operationId": "arazzoMermaid",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OasInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "OK",
            "headers": {
              "API-Version": {
                "description": "De API-versie van de response",
                "explode": false,
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "apiKey": [],
            "clientCredentials": []
          }
        ],
        "summary": "Arazzo Mermaid (POST)",
        "tags": [
          "Tools"
        ]
      }
    },
    "/v1/auth/clients": {
      "post": {
        "description": "Maak een client aan via de admin API. Body bevat Email.",
        "operationId": "untrustedClient",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UntrustedClientInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeycloakClientResult"
                }
              }
            },
            "description": "OK",
            "headers": {
              "API-Version": {
                "description": "De API-versie van de response",
                "explode": false,
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "apiKey": [],
            "clientCredentials": []
          }
        ],
        "summary": "Maak client (POST)",
        "tags": [
          "Tools"
        ]
      }
    },
    "/v1/oas/convert": {
      "post": {
        "description": "Converteert OpenAPI naar de laatst ondersteunde versie (standaard 3.1). Meegegeven targetVersion (3.0 of 3.1) bepaalt het doel. Body: { oasUrl } of { oasBody } (stringified JSON of YAML).",
        "operationId": "ConvertOAS",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OasInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "API-Version": {
                "description": "De API-versie van de response",
                "explode": false,
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "apiKey": [],
            "clientCredentials": []
          }
        ],
        "summary": "Converteer OpenAPI 3.0/3.1",
        "tags": [
          "Tools"
        ]
      }
    },
    "/v1/oas/bundle": {
      "post": {
        "description": "Bundelt een OpenAPI specificatie en lost externe verwijzingen op. Body: { oasUrl } of { oasBody }.",
        "operationId": "bundleOAS",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OasInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "API-Version": {
                "description": "De API-versie van de response",
                "explode": false,
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "apiKey": [],
            "clientCredentials": []
          }
        ],
        "summary": "Bundle OpenAPI",
        "tags": [
          "Tools"
        ]
      }
    },
    "/v1/oas/generate": {
      "post": {
        "description": "Genereert een boilerplate OpenAPI specificatie op basis van JSON-invoer. Body: { oasUrl } of { oasBody } (stringified JSON).",
        "operationId": "GenerateOAS",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OasInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "API-Version": {
                "description": "De API-versie van de response",
                "explode": false,
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "apiKey": [],
            "clientCredentials": []
          }
        ],
        "summary": "Generate OpenAPI",
        "tags": [
          "Tools"
        ]
      }
    },
    "/v1/oas/validate": {
      "post": {
        "description": "Valideert een OpenAPI specificatie met de DON ADR ruleset. Standaard wordt ruleset 2.1 gebruikt; geef targetVersion \"2.0\", \"2.1\", \"2.2\" of \"draft\" mee om een versie te kiezen. Body: { oasUrl } of { oasBody } (stringified JSON of YAML).",
        "operationId": "validatorOpenAPIPost",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LintResult"
                }
              }
            },
            "description": "OK",
            "headers": {
              "API-Version": {
                "description": "De API-versie van de response",
                "explode": false,
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "apiKey": [],
            "clientCredentials": []
          }
        ],
        "summary": "Validate OpenAPI (POST)",
        "tags": [
          "Tools"
        ]
      }
    },
    "/v1/oas/postman": {
      "post": {
        "description": "Converteert OpenAPI naar Postman Collection JSON. Body: { oasUrl } of { oasBody } (stringified JSON of YAML).",
        "operationId": "CreatePostmanCollection",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OasInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "API-Version": {
                "description": "De API-versie van de response",
                "explode": false,
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "security": [
          {
            "apiKey": [],
            "clientCredentials": []
          }
        ],
        "summary": "Maak Postman-collectie (POST)",
        "tags": [
          "Tools"
        ]
      }
    }
  },
  "components": {
    "headers": {
      "API-Version": {
        "description": "De API-versie van de response",
        "explode": false,
        "schema": {
          "example": "1.0.0",
          "type": "string"
        },
        "style": "simple"
      }
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "headers": {
          "API-Version": { "$ref": "#/components/headers/API-Version" }
        },
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
      "404": {
        "description": "Resource does not exist",
        "headers": {
          "API-Version": { "$ref": "#/components/headers/API-Version" }
        },
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
      "409": {
        "description": "Conflict",
        "headers": {
          "API-Version": { "$ref": "#/components/headers/API-Version" }
        },
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "headers": {
          "API-Version": { "$ref": "#/components/headers/API-Version" }
        },
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      }
    },
    "schemas": {
      "Problem": {
        "description": "RFC 7807 Problem Details for HTTP APIs.",
        "type": "object",
        "required": ["type", "title", "status"],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "description": "URI reference identifying the problem type."
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the problem."
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code."
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence."
          },
          "instance": {
            "type": "string",
            "description": "URI reference identifying the specific occurrence."
          },
          "errors": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProblemError" },
            "description": "Field-level errors, when applicable."
          }
        }
      },
      "ProblemError": {
        "type": "object",
        "required": ["detail", "pointer"],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable description of this specific error."
          },
          "pointer": {
            "type": "string",
            "description": "JSON Pointer (RFC 6901) identifying the offending field, prefixed with '#'."
          }
        }
      },
      "OasInput": {
        "description": "Geef oasBody (stringified spec) of oasUrl (URL naar de spec). Voor Arazzo endpoints zijn ook arazzoBody / arazzoUrl toegestaan.",
        "example": {
          "oasUrl": "https://example.com/openapi.yaml",
          "targetVersion": "3.1"
        },
        "properties": {
          "oasBody": {
            "type": "string",
            "minLength": 1
          },
          "oasUrl": {
            "type": "string",
            "format": "uri"
          },
          "arazzoBody": {
            "type": "string",
            "minLength": 1
          },
          "arazzoUrl": {
            "type": "string",
            "format": "uri"
          },
          "targetVersion": {
            "description": "Doelversie. Voor conversie: 3.0 of 3.1. Voor validatie: 2.0 of 2.1.",
            "type": "string"
          }
        },
        "anyOf": [
          { "required": ["oasBody"] },
          { "required": ["oasUrl"] },
          { "required": ["arazzoBody"] },
          { "required": ["arazzoUrl"] }
        ],
        "type": "object"
      },
      "ValidateInput": {
        "description": "Body voor de OpenAPI validate endpoint. targetVersion kiest de ADR ruleset (default 2.1).",
        "allOf": [
          { "$ref": "#/components/schemas/OasInput" },
          {
            "type": "object",
            "properties": {
              "targetVersion": {
                "type": "string",
                "enum": ["2.0", "2.1", "2.2", "draft"]
              }
            }
          }
        ]
      },
      "KeycloakClientResult": {
        "example": {
          "apiKey": "apiKey"
        },
        "properties": {
          "apiKey": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "LintMessage": {
        "example": {
          "severity": "severity",
          "createdAt": "2000-01-23T04:56:07.000Z",
          "code": "code",
          "id": "id",
          "infos": [
            {
              "path": "path",
              "lintMessageId": "lintMessageId",
              "id": "id",
              "message": "message"
            },
            {
              "path": "path",
              "lintMessageId": "lintMessageId",
              "id": "id",
              "message": "message"
            }
          ]
        },
        "properties": {
          "code": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "infos": {
            "items": {
              "$ref": "#/components/schemas/LintMessageInfo"
            },
            "type": "array"
          },
          "severity": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "LintMessageInfo": {
        "example": {
          "path": "path",
          "lintMessageId": "lintMessageId",
          "id": "id",
          "message": "message"
        },
        "properties": {
          "id": {
            "type": "string"
          },
          "lintMessageId": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "LintResult": {
        "example": {
          "createdAt": "2000-01-23T04:56:07.000Z",
          "score": 6,
          "failures": 0,
          "successes": true,
          "rulesetVersion": "2.1",
          "messages": [
            {
              "severity": "severity",
              "createdAt": "2000-01-23T04:56:07.000Z",
              "code": "code",
              "id": "id",
              "infos": [
                {
                  "path": "path",
                  "lintMessageId": "lintMessageId",
                  "id": "id",
                  "message": "message"
                },
                {
                  "path": "path",
                  "lintMessageId": "lintMessageId",
                  "id": "id",
                  "message": "message"
                }
              ]
            },
            {
              "severity": "severity",
              "createdAt": "2000-01-23T04:56:07.000Z",
              "code": "code",
              "id": "id",
              "infos": [
                {
                  "path": "path",
                  "lintMessageId": "lintMessageId",
                  "id": "id",
                  "message": "message"
                },
                {
                  "path": "path",
                  "lintMessageId": "lintMessageId",
                  "id": "id",
                  "message": "message"
                }
              ]
            }
          ],
          "id": "id",
          "apiId": "apiId"
        },
        "properties": {
          "apiId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "failures": {
            "format": "int32",
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/LintMessage"
            },
            "type": "array"
          },
          "score": {
            "format": "int32",
            "type": "integer"
          },
          "successes": {
            "type": "boolean"
          },
          "rulesetVersion": {
            "description": "De gebruikte ruleset-versie voor validatie.",
            "type": "string"
          }
        },
        "type": "object"
      },
      "UntrustedClientInput": {
        "example": {
          "email": "user@example.com"
        },
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": ["email"],
        "type": "object"
      }
    },
    "securitySchemes": {
      "apiKey": {
        "in": "header",
        "name": "X-Api-Key",
        "type": "apiKey",
        "description": "Read-only API keys for tools endpoints can be requested at https://apis.developer.overheid.nl/apis/key-aanvragen."
      },
      "clientCredentials": {
        "flows": {
          "clientCredentials": {
            "scopes": {
              "tools": "Access to tools endpoints"
            },
            "tokenUrl": "https://auth.developer.overheid.nl/realms/don/protocol/openid-connect/token"
          }
        },
        "type": "oauth2"
      }
    }
  }
}